Crate subspace_runtime

Modules

api

Structs

AvgBlockspaceUsageNumBlocks

BlockAuthoringDelay

BlockHashCount

BlockHashProvider

BlockSlot

BlockSlotCount

BlockTreePruningDepth

BlockchainHistorySize

BundleLongevity

ChannelFeeModel

ChannelInitReservePortion

ChannelReserveFee

ConfirmationDepthK

CooloffPeriod

CouncilMotionDuration

CreditSupply

DemocracySlash

DisablePallets

Disable specific pallets.

DomainInstantiationDeposit

DomainRegistration

DomainRuntimeUpgradeDelay

Runtime upgrade is delayed for 1 day at 6 sec block time.

DomainTxRangeAdjustmentInterval

DomainsPalletId

DynamicCostOfStorage

EnactmentPeriod

EraDuration

ExistentialDeposit

ExpectedVotesPerBlock

FastTrackVotingPeriod

HoldIdentifierWrapper

InitialDomainTxRange

LaunchPeriod

MaxDomainBlockSize

Use the consensus chain’s Normal extrinsics block size limit as the domain block size limit

MaxDomainBlockWeight

Use the consensus chain’s Normal extrinsics block weight limit as the domain block weight limit

MaxDomainNameLength

MaxInitialDomainAccounts

MaxPendingStakingOperation

MaxProposalWeight

MaximumReceiptDrift

MaximumSchedulerWeight

MinInitialDomainAccountBalance

MinNominatorStake

Minimum nominator stake to nominate and operator.

MinOperatorStake

Minimum operator stake to become an operator.

MinSectorLifetime

MmrProofVerifier

MmrRootHashCount

NoPreimagePostponement

OnChainRewards

OnXDMRewards

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

PotEntropyInjectionDelay

PotEntropyInjectionInterval

PotEntropyInjectionLookbackDepth

PreImageHoldReason

PreimageBaseDeposit

PreimageByteDeposit

ProposerTaxOnVotes

RecentHistoryFraction

RecentSegments

RewardPoint

Runtime

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeGenesisConfig

RuntimeOrigin

The runtime origin type representing the origin of a call.

SelfChainId

SessionKeys

ShouldAdjustSolutionRange

SlotProbability

StakeEpochDuration

StakeWithdrawalLockingPeriod

StorageKeyProvider

StorageKeys

SubspaceBlockLength

We allow for 3.75 MiB for Normal extrinsic with 5 MiB maximum block length.

SubspaceBlockWeights

We allow for 2 seconds of compute with a 6 second average block time.

TotalSpacePledged

TransactionWeightFee

TransporterEndpointId

TreasuryAccount

Version

VoteLockingPeriod

VotingPeriod

WithdrawalLimit

Enums

AllowAuthoringBy

EnableRewardsAt

When to enable block/vote rewards

OriginCaller

RuntimeCall

The aggregated runtime call type.

RuntimeError

RuntimeEvent

RuntimeFreezeReason

A reason for placing a freeze on funds.

RuntimeHoldReason

A reason for placing a hold on funds.

RuntimeLockId

An identifier for each lock placed on funds.

RuntimeSlashReason

A reason for slashing funds.

RuntimeTask

An aggregation of all Task enums across all pallets included in the current runtime.

Constants

RUNTIME_API_VERSIONS

VERSION

WASM_BINARY

WASM_BINARY_BLOATY

Type Aliases

Address

The address format for describing accounts.

AllCouncil

AllPalletsWithSystem

All pallets included in the runtime as a nested tuple of types.

AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

Balances

BalancesConfig

Block

Block type as expected by this runtime.

Council

CouncilCollective

CouncilConfig

Democracy

DemocracyConfig

Domains

DomainsConfig

EnsureRootOr

Executive

Executive: handles dispatch to the various modules.

HalfCouncil

Header

Block header type as expected by this runtime.

Messenger

Mmr

Preimage

Rewards

RewardsConfig

RuntimeConfigs

RuntimeConfigsConfig

SS58Prefix

Scheduler

SignedExtra

The SignedExtension to the basic transaction logic.

Subspace

SubspaceConfig

SubspaceMmr

Sudo

SudoConfig

System

SystemConfig

Timestamp

TransactionFees

TransactionPayment

TransactionPaymentConfig

Transporter

TwoThirdsCouncil

UncheckedExtrinsic

Unchecked extrinsic type as expected by this runtime.

Utility