subspace_runtime

Type Alias AllPalletsWithSystem

pub type AllPalletsWithSystem = (System, Timestamp, Subspace, Rewards, Balances, TransactionFees, TransactionPayment, Utility, Domains, RuntimeConfigs, Mmr, SubspaceMmr, Messenger, Transporter, Scheduler, Council, Democracy, Preimage, Multisig, Sudo);

All pallets included in the runtime as a nested tuple of types.