Type Alias subspace_runtime::RuntimeConfigsConfig

pub type RuntimeConfigsConfig = GenesisConfig<Runtime>;

Aliased Type

struct RuntimeConfigsConfig {
    pub enable_domains: bool,
    pub enable_dynamic_cost_of_storage: bool,
    pub enable_balance_transfers: bool,
    pub enable_non_root_calls: bool,
    pub confirmation_depth_k: u32,
}

Fields
enable_domains: bool

Whether to enable domains

enable_dynamic_cost_of_storage: bool

Whether to enable dynamic cost of storage (if false cost per byte is equal to 1)

enable_balance_transfers: bool

Whether to enable balance transfers

enable_non_root_calls: bool

Whether to enable calls from non-root account

confirmation_depth_k: u32

Confirmation depth k to use in the archiving process