subspace_runtime

Type Alias RuntimeConfigsConfig

pub type RuntimeConfigsConfig = GenesisConfig<Runtime>;

Aliased Type

pub struct RuntimeConfigsConfig {
    pub enable_domains: bool,
    pub enable_dynamic_cost_of_storage: bool,
    pub enable_balance_transfers: bool,
    pub confirmation_depth_k: u32,
    pub council_democracy_config_params: CouncilDemocracyConfigParams<u32>,
    pub domain_block_pruning_depth: u32,
    pub staking_withdrawal_period: u32,
}

Fields
enable_domains: bool

Whether to enable domains

enable_dynamic_cost_of_storage: bool

Whether to enable dynamic cost of storage (if false cost per byte is equal to 1)

enable_balance_transfers: bool

Whether to enable balance transfers

confirmation_depth_k: u32

Confirmation depth k to use in the archiving process

council_democracy_config_params: CouncilDemocracyConfigParams<u32>

Council and democracy config params.

domain_block_pruning_depth: u32

Domain block pruning depth.

staking_withdrawal_period: u32

Domain nominator’s staking withdrawal period.